Cleanup general functions and put in utils.py
This commit moves same functions multiplied in different files into one
file (utils.py).
